+++ packages/mock/src/client/mock.ts
@@ -1,16 +1,15 @@
 import { Middleware } from 'fatcher';
 import { MockOptions } from './interfaces';
 import { parser } from '../parser';
-import { swChecker } from './swChecker';
-
-const MOCK_HEADER_KEY = 'x-fatcher-mock';
+import { checker } from './checker';
+import { MOCK_HEADER_KEY } from '../utils';
 
 export function mock(options: MockOptions = {}): Middleware {
     const { enabled = process.env.NODE_ENV !== 'production', schema } = options;
 
     return {
         name: 'fatcher-middleware-mock',
-        presets: [swChecker()],
+        presets: [checker()],
         async use(context, next) {
             if (enabled) {
                 if (schema) {
