+++ packages/mock/src/utils/index.ts
@@ -1,0 +1,1 @@
+export const MOCK_HEADER_KEY = 'x-fatcher-mock';
